+++ src/components/Main/Metrics/MetricsEmployee/MetricsEmployee.jsx
@@ -1,10 +1,15 @@
 import { useEffect, useState } from "react";
+import Carbon from "./Carbon/Carbon.jsx";
 import Employee from "./Employee/Employee.jsx";
 import Office from "./Office/Office.jsx";
-import Team from "./Team/Team.jsx";
+import BarChartComponent from "./BarChart/BarChart.jsx";
 import Workstation from "./Workstation/Workstation.jsx";
+import PieChartComponent from "./PieChart/PieChart.jsx";
 import { $MetricsEmployee } from "./MetricsEmployee.styles.jsx";
-import { apiCallToGetEmployeeMetric } from "../../../../api/Api.jsx";
+import {
+  apiCallToGetEmployeeMetric,
+  apiCallToGetEmployeeCarbonFootprint,
+} from "../../../../api/Api.jsx";
 
 export default function MetricsEmployee() {
   const [employeeMetric, setEmployeeMetric] = useState({});
@@ -13,6 +18,8 @@ export default function MetricsEmployee() {
     const metrics = apiCallToGetEmployeeMetric(0);
     setEmployeeMetric((_prevMetrics) => metrics);
   }, [employeeMetric]);
+
+  const carbonFootprint = apiCallToGetEmployeeCarbonFootprint(1);
 
   const {
     id,
@@ -27,22 +34,22 @@ export default function MetricsEmployee() {
     workstation_id,
   } = employeeMetric;
   const employee = {
-    id,
     corporate_key,
     email,
     name,
     surname,
     role,
-    carbon_limit,
     location,
   };
-  const { team } = employeeMetric;
 
   return (
     <$MetricsEmployee>
       <Employee employee={employee} />
       <Office office={office_id} />
       <Workstation workstation={workstation_id} />
+      <Carbon employeeId={id} carbonFootprint={carbonFootprint} carbonLimit={carbon_limit} />
+      <BarChartComponent carbonFootprint={carbonFootprint} carbonLimit={carbon_limit} />
+      <PieChartComponent carbonFootprint={carbonFootprint} />
     </$MetricsEmployee>
   );
 }