+++ src/components/Main/Metrics/MetricsEmployee/Team/Team.jsx
@@ -1,7 +1,0 @@
-import { $Team } from "./Team.styles.jsx";
-
-export default function Team() {
-    return (
-        <$Team>team</$Team>
-    );
-}